Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    int a,b;
-    scanf("%d",&a);
+    float a,b;
+    scanf("%f",&a);
     b=3.14*a;
-    printf("Area: %d",b);
+    printf("Area: %f",b);
 }
